Rename smart_home_onf.hocon to smart_home.hocon
