+++ coded_tools/smart_home/kitchen_lights_switch.py
@@ -7,16 +7,16 @@
 # Purchase of a commercial license is mandatory for any use of the
 # neuro-san-demos SDK Software in commercial settings.
 #
-from coded_tools.smart_home_onf.lights_switch import LightsSwitch
+from coded_tools.smart_home.lights_switch import LightsSwitch
 
 
-class LivingRoomLightsSwitch(LightsSwitch):
+class KitchenLightsSwitch(LightsSwitch):
     """
     CodedTool implementation that calls an API to turn lights on or off.
     """
 
     def __init__(self):
         """
-        Constructs a switch for living room lights.
+        Constructs a switch for kitchen lights.
         """
-        super().__init__("Living room")
+        super().__init__("Kitchen")
