+++ registries/manifest.hocon
@@ -16,7 +16,6 @@
     "airline_policy.hocon": true,
     "advanced_calculator.hocon": true,
     "smart_home.hocon": true,
-    "smart_home_onf.hocon": true,
     "smart_home_include.hocon": true,
     "agent_network_designer.hocon": true,
     "agent_network_generator.hocon" : true,
